Add files via upload
This is a user profile page that contains the user's name, surname, DOB, their phone number, email address and if they want to change their password. Also, if the user forgets their password, they can simply click on the forgot password and it will take them to the forgot password page, where the user will input their email address and the link to the reset password will be sent to their email.